requirements.txt
numpy>=1.23.2
opencv_python>=4.6.0.66
requests>=2.28.1
ultralytics
pillow